[feature #161432039] Admin can create a user and user gets a token on login
